refactor: authenticated room into player component
